requirements.txt
black==23.7.0
ipykernel==6.25.1
ipython==8.14.0
isort==5.13.2
flake8==6.1.0
polars==0.20.3
pytorch-lightning==2.1.3
pandas==2.0.3
seaborn==0.12.2
matplotlib==3.7.4
PyWavelets==1.4.1
librosa==0.10.1
pyarrow==11.0.0
albumentations==1.3.1
timm==0.9.12
wandb==0.16.2